test: keep cold-start coverage maintainable
