+++ pkg/transports/cli/root_factory.go
@@ -495,9 +495,18 @@ func prepareRunSystemInitialization(cmd *cobra.Command, cfg *runcli.RunConfig, o
 // named/bootstrap, continuous, and real-worker invocations retain the normal
 // initialization boundary.
 func deferBatchSystemInitialization(cmd *cobra.Command, cfg runcli.RunConfig) bool {
-	if cmd != nil && (cmd.Flags().Changed("dir") || cmd.Flags().Changed("factory") || cmd.Flags().Changed("named")) {
+	if runFactorySelectionWasExplicit(cmd) {
 		return false
 	}
+	return isFiniteMockNoRecordBatch(cfg)
+}
+
+func runFactorySelectionWasExplicit(cmd *cobra.Command) bool {
+	return cmd != nil &&
+		(cmd.Flags().Changed("dir") || cmd.Flags().Changed("factory") || cmd.Flags().Changed("named"))
+}
+
+func isFiniteMockNoRecordBatch(cfg runcli.RunConfig) bool {
 	return strings.TrimSpace(cfg.WorkFile) != "" &&
 		!cfg.Continuously &&
 		cfg.MockWorkersEnabled &&
